refactor(admin): 이미지 업로드 로직 분리 - 모달에서 쓰기 위해 상위 컴포넌트로 분리
